CMS7-5479: added link picker for the content root of a channel.
The ExtJS xtype 'linkpicker' renders an ExtJS widget that shows the current path and allows the user to pick a new path via the existing Wicket link picker dialog. The Wicket dialog is rendered to a placeholder div that is wrapped in the ExtLinkPickerContainer. Clicking the ExtJS button fires a 'pick' event on the container, which then renders the link picker. The picked path is returned to the original component via a 'picked' event generated by the ExtLinkPicker Ajax response.
 
 Also refactored the code base a bit: moved all widgets to a separate package 'widgets', and changed the visibility of some helper classes to package-private.