CMS7-5518: work-in-progress
Split editor into a container (iframe manager) and the UI part.
First futures have been introduced for retrieving the iframe resources in
parallel.

Page editor is not functional, we still need to find a better way of capturing
the dependencies.  Plain events don't cut it, it gets much too complex.