CMS7-5556 Channel Manager breadcrumb
The breadcrumb implementation is based on a toolbar. Instead of removing / disabling toolbar-items this is a fallback implementation to workaround a rendering issue of ExtJS. Ext is not handling removal and hiding of toolbar-items properly which results in the whole toolbar gets hidden / removed. The issue seems related to the workaround which is used for switching the cards of the RootPanel.