CMS7-5556: use 'cards' (ext panel + title property) for breadcrumb
The breadcrumb bar now observes title changes, causing the breadcrumb item to update it's text when the channel changes.  The channel manager and hst config editor names are now retrieved by wicket, to make it easier to localize.
  
The channel name has been removed from the "workflow" toolbar.

Typo in the PageEditor was fixed to test whether editor already exists.

 thanks jannis for review