CMS7-5569 channel properties window in channel editor
The properties side panel in the channel overview has been removed and added as a window inside the channel editor (page editor). A new toolbar button is showing / closing the channel properties window.
 
The channel manager service is now using the channel id in combination with the subMountPath, which is a result of the channel store moved to the RootPanel from where it is shared wrapped in a future with the page editor and the channel grid panel.