CMS7-5582: injected anchor element after the hst meta data comment
The dom is queried for comment nodes which get parsed and if succeeded a anchor element gets inserted after the comment. The comment is removed to indicate the process status.
There are two variants implemented to get all comment nodes the first one is done by xpath (fast), the fallback should be done by dom traversal or via document.getElementsByTagName('*') [1].

[1] http://ejohn.org/blog/getelementsbyclassname-speed-comparison/