CMS7-5663: only show the 'add channel' button (and toolbar) when blueprints are available.
When the 'add channel' button is not functional (either because no blueprints are available, or the currect user is not allowed to add channels) the 'add channel' button will not be shown. Since this results in an empty toolbar, the toolbar itself is also not shown anymore to gain some screen space.