CMS7-5886: Support channel fields as placeholder names in the icon paths, styling improvements
The icon paths for the locale and the icon of the channel have been separated, in the path the channel fields can be used as placeholders e. g. ${type}. There could be even a more detailed implementation to support cascaded fallbacks of icons.
The switching view buttons have been layouted according to the mock. The 'type' / 'region' switch in the channel icon overview panel will stay as is for now.