CMS7-6022: backport, always reload the toolkit
The toolkit used to be loaded only once, and reused when the same mount was edited again. However, when the hst:catalog items in the toolkit are part of the channel's configuration itself (as is the case in, for example, the current archetype), their UUIDs will change whenever a new preview configuration is created. In that case, the cached toolkit contains UUIDs that no longer exist, causing an ItemNotFoundException when these items are dragged into a container. To prevent this from happening, we now always reload the toolkit.