CMS7-6072: enable the document picker in the template composer
Component properties can now be annotated with the JcrPath annotation, which will render a linkpicker widget that opens the good old document picker. When the JcrPath is relative, the root level of the navigator will be set to the content root of the current channel, so you cannot select documents that do no belong to your own channel.
 
Channel properties can now also be annotated with relative JcrPaths, which will interpret the path as one relative to the content root of the channel (the same way as relative paths for component properties do).