CMS7-6230 Use hst-preview content as context to initialize the Hst Config Editor
The HST Config Editor is using the live content facet to render the tree view and documents in the link picker and so not showing unpublished documents. The default has been changed to pass the preview content facet path, so it can be used to create the context for the HST Config Editor. The improvement can be turned off via the plugin configuration property initializeHstConfigEditorWithPreviewContext.