CMS7-6215 New channel groups, added due to adding a channel are not getting rendered
Each channel group is rendered separate data view. If a channel gets added with a previously not rendered group, the channel and group are not visible until a hard refresh is done. The new approach is unifying the individual views into one data view, for which the channel data gets aggregated to suit the view's XTemplate. Reloading of the store data is calling refresh, which registers the click listeners for the group handlers.