CMS7-6587 the component blocks in the template composer become distorted when selecting variant
The synchronization of the overlay happens before the actual rendering of the component succeeded. Unfortunately browsers don't provide an event mechanism for rendering of elements. The element of the component gets actively polled for 5 seconds to check its rendering. When the rendering succeeds or the polling times out the callback gets called.