CMS7-6896: update the iframe when the window is resized
Removed the 'iframe.purgeListeners' call, which removed the 'resize' listener registered by iframe itself so no more 'resize' messages were sent to the iframe.
Also added throttling of 'resize' messages to reduce the amount of messages sent to the iframe and to give the page inside the iframe time to reflow.