CMS7-7089: load device properties in the channel store
Pass them to the device manager when creating the toolbar plugin.
Improve device record lookup logic to deal with invalid or unavailable devices.

The list of available devices is now easily extended by adding a devices/ folder to a project.
Note that it is still possible to create other DeviceService implementations that would e.g. fetch their skins from the repository.

Some renaming to make the class names consistent.

The mask on the iframe has been restored to block user interaction after navigation, when the template composer is updating.

Thanks to @mdenburger for pairing!