CMS7-7120: backport CMS7 6849 - allow pages in the template composer to return an error status
When a page returns an HTTP error status, we still try to initialize the page context. This allows, for example, editing of '404' pages. Only when a page returns a status code of 500 or higher, or the response is missing vital information to initialize the page context, we show an error message 'This page returned an error, going back.'. After clicking OK the previous page is then opened again. If no previous page is available, we go back to the channel manager overview.