CMS7-7128: backport - make channel preview visible for authors again
For non-authors, showing the iframe happens (surprisingly) when toolbar plugins are inserted 
(the 'add' events causes the toolbar owner - the iframe - to show itself). 

For authors, adding toolbar plugins caused warnings since the components referenced for the 
position of the plugin ('before' or 'after' component X) did not exist. We now only add
 toolbar plugins when the default toolbar items have been added too, which avoids the warnings.
 
When nothing is added to the toolbar (i.e. for authors) we have to make sure the iframe 
gets shown anyway. The existing odd afterrender code of the toolbar did not actually do anything, 
so I replaced it by a simple 'show()' call for the iframe.

I also made sure that the 'mask' classes are always removed from the template composer instance. 
Masking may put these in there when the iframe is not yet visible, but unmasking always happens 
when the iframe is visible so the mask classes never got removed again. They will now always be 
removed.