CMS7-7231: backport CMS7 7223 - show changes notification to authors
The changes notification is positioned relative to the iframe panel toolbar, 
which may not yet be positioned correctly when the toolbar is empty (which is
currently always the case for authors).
 
For editors and admins, the toolbar will always include the 'device skin' combo, 
which sets a CSS class on the iframe panel element, which has the side effect that
the toolbar is positioned correctly within the hidden iframe. We should not rely 
on this side effect, so we now position the notification after the iframe panel 
itself has been shown (and the toolbar in it has been positioned correctly).