CMS7-7410: backport CMS7 7403 - auto-adjust component properties window height such that both the current component properties editor and all variant tabs are fully visible
Introduced a new API method for component properties editors, 'syncVisibleHeight', which is responsible for calculating the visible height 
of the editor and firing a 'visibleHeightChanged' event whenever that height changed. The component properties panel also reports its
visible height this way, which is the maximum of the height needed for the variant tabs and the height needed for the editor. 
The window adjusts its size accordingly, and also checks that its height does not become bigger than the height of the page editor itself.
           
The 'visible height' concept was introduced because the component height reported by ExtJS is not always correct (e.g. it does not take 
padding and margins of box components into account).