CMS7-7671: add dummy Angular app + communication with container
The container provided the ID of the Ext IFramePanel component that contains the app in a query parameter. The Angular service 'hippo.channelManager.Container' parses this parameter and reads the app configuration from the 'pageManagementConfig' property of the IFrameComponent. Communication between Angular and the container happens via the message busses hostToIFrame and iframeToHost of the IFramePanel. When the Angular app is not loaded in an iframe, communication will be ignored.