CMS7-7862: prevent long warnings from showing the bottom of the confirmation dialog
The confirmation dialog gets a negative 'top' property, which is animated in capable 
browsers when it's set to '0'. However, the proper negative top value depends on the
height of the confirmation dialog, which can change depending on the message shown.

We now limit the height of the message to 2 lines, so the confirmation dialog does
not grow too high when the message changes via auto-binding.