CMS7-8034 Gracefully shutdown the executor service
By letting Wicket manage the lifecycle of the RestProxyServicesManager we can make sure that
the executor service is shutdown when the Wicket application shuts down. This way the 
application container (tomcat e.g.) will to be blocked by active threads. In addition, the
threads used by the executor service are now all daemon threads too although this is not
really necessary.