CHANNELMGR-50 Remove hippo-plugins dependency
The channel manager currently depends on hippo-plugins, but it is not used in the source code.
Because hippo-plugins.js depends on js-detect-ie, the channel manager depends on it too.
Downloading js-detect requires a git client, which might not be installed by all users.
So, since we do not use hippo-plugins in the channel manager, we can safely remove it.
This will make it possible to build the channel manager without installing a git client.