CMS7-8594: use delayed tasks for breadcrumb animation
The breadcrumb should be hidden after the browse perspective is shown, so the clicked tab does not 'move to the left' immediately. The show and hide animation therefore became separate tasks. The hide animation is delayed for 0.5 seconds to give the browse perspective time to appear. Each animation now cancels the previous one, so quickly switching between the browse and channel manager perspective does not result in a crazy breadcrumb trying to keep up. Switching perspectives during the animation now triggers the other animation smoothly.