CHANNELMGR-97 prefix filters and services with namespace
Since the filters and services are now open for use in
other AngularJS apps the services and filters should be
name-spaced to prevent name collisions.

TODO: tooltipShow directive. Besides the channel manager,
this directive is also used in frontend of targeting, so I
can only commit it if I modify it in targeting too.