CHANNELMGR-197: update dirty state on key up without triggering component preview
When editing text or number fields, the record value and dirty state are updated directly so the 'save' button also becomes enabled directly. To avoid too many calls, keyup does not re-render a preview version of the component (that only happens when hitting ENTER, blurring the field or changing it otherwise).