CHANNELMGR-201: do not wait for re-rendering of an empty component
This applies to hiding a component via the 'Hide' option in the component properties box. The component rendering response is then empty, which used to trigger the 5-second wait (since the component's height is that of an empty component). That's not necessary since we know we're rendering an empty component, so we now return immediately.