CHANNELMGR-24: save string version of date when selected and prevent manual editing
The value of a datefield is Date, not a string. The 'select' and 'change' event of a datefield therefore stored the Date.toString value instead of the parsed date. We now parse the date before saving it.
 
The 'editable = false' prevents manual editing of the date as a string, and always shows the picker. This prevents weird behavior when someone enters odd date values.

Also prevented updating a text field with ENTER when its not valid.