CHANNELMGR-179: open 'edit menu' dialog when 'edit menu' button behind container item overlay is 'clicked'
The click will go to the container item overlay, so we manually test whether the same point in the viewport 
contains an 'edit menu' button. If so, we don't show the component properties but 'click' the edit menu 
button to open the 'edit menu' dialog.
 
The points from the click event are in page coordinates, but document.elementFromPoints wants viewport
coordinates. Hence we subtract the scroll offset in the iframe from the page coordinates so the
edit menu button can also be clicked when the page in the iframe is scrolled.