CHANNELMGR-336: Compute the iframeUrl instead of using the channel.url property
The preview url of a channel should be computed using contextPath, cmsPreviewPrefix, mountPath and an optional channelPath (e.g. a document in a channel). 

I removed the contextPath and CmsPreviewPrefix parameters from the ChannelEditor as these are already known on the client by means of the (EXT) ChannelStore. Theoretically a channelId should be enough information to render a channel. Subsequently a channelId and a channelPath should be enough information to open a specific page in a specific channel. For this I altered the IChannelManagerService.