CHANNELMGR-336: Replace ES6 promise with $q
Using ES6 promises in Angular 1 code is a bad idea since $q promises are tied to the digest cycle and ES6 promises don't provide a hook for Angular to do that. This allowed me to remove the 'done' pattern from the unit tests and simplify most of the test code.