CHANNELMGR-372 move channel editor to separate package
Don't load the template composer source code anymore. All remaining
ExtJs code for the channel editor should move to the channel
editor, so the 'templatecomposer' package can be deleted entirely when
the channel editor is ready.

The plugins already moved to the channel editor (we will need all of
them in the channel properties editor dialog).