CHANNELMGR-463 create preview HST configuration when clicking 'Edit'
The parsed page meta-data is now stored in the PageMetaDataService.
When there is no preview configuration yet (as indicated in the page
meta-data) we create it. The 'Edit' button is disabled while the
preview configuration is being created.

Also added a unit test for the show-component-properties event. Since
the run() function of the app also generates an event
('reload-channel') tests now only spy on the mocked event buses when
they expect things on them so they only have to expect the events
they generate themselves.

Also rephrased the test description in pageStructureService.spec.js
to use active tense.