CHANNELMGR-458 Sync iframe height after the Angular app his finished loading
Let the overlaySync register it's own onLoad handler on the iframe element. In it, we sync the iframe height, start observing DOM mutations, register a resize listener on the main window and register an unload handler. Previously it could happen that the iframe height was being synced while the app was still loading.