CHANNELMGR-495 destroy dragula promise when navigating to other page
The promise indicates that dragula has been injected into the iframe.
Before another page is loaded, the promise should be destroyed so
dragula will be re-injected into the next page.

Since the beforeunload call will be executed only once, the event
handler should be registered with 'one' instead of 'on'. The next
page load will register the beforeunload handler again for the next
iframe window.