CHANNELMGR-468 style dragged element, show containers while dragging
The dragula mirror is rendered in the app so it can be positioned
over the container lines. In order to keep the mirror styled correctly,
the computed style of all elements in the cloned element is copied
recursively to the elements in the mirror, except for the properties
that matter for dragula:

 - opacity should be set be dragula
 - pointer-event should be none so the mirror does not obscure
   drag events
 - user-select should be 'none' (including vendor prefixes)

In addition, the root element of the mirror should get its borders,
box-shadow etc. from the overlay style.