CHANNELMGR-470 Fix an issue in Firefox where the height of the iframe is incorrect upon load
The throttle function executes the callback and then starts blocking incoming calls. This means that the state of the callback owner can be out of sync, as happens in Firefox when using the throttleService to throttle MutationObserver callbacks. Moving the callback call into the timeout handler ensures the state is correct, but might trigger a second callback. This might need a different approach, like maybe re-using the throttle from lodash.