CHANNELMGR-468 preserve width and height of mirror set by Dragula
Copying the computed style to the mirror overwrote the 'width' and
'height' properties that Dragula sets before emitting the 'cloned'
event. The DomService now preserves all style attributes of the
target element that are not copied from the source element, and the
DragDropService does not copy the width and height of the source
element Dragula's values are no longer overwritten.