CHANNELMGR-533 Update channel config when channel-loaded is fired
Extjs loads the channel config store on page load, if by chance
the repository changes and the page is not reloaded the angularjs
app would be working with an out of date channel config.
Current solution gets the channel config straight from the backend
whenever a channel-load event is fired for example when selecting
a channel in the channel overview.