CHANNELMGR-542 improve performance of the drag mirror creation in IE11
The main trick is to detach the mirror's children from the DOM while
copying the computed style of the original element to it. This prevents
reflows, which speeds up the code about 4x in IE11.

It's not possible to detach the mirror element itself because Dragula
keeps a reference to it; detaching and re-adding the mirror breaks
that reference. That's why we only detach all children.

Also did a few micro optimizations in _copyComputedStyle:

 - append properties to an array that's eventually joined
   instead of doing string concatenation all the time
 - cache the fromComputedStyle.length property
 - don't copy empty string values