CHANNELMGR-552: reloaded the channel-store whenever switching-channel event is sent from AngularJS app
When editing a channel at the first time, the channel-store in ExtJs is outdated because the 'edit-channel' request creating the '-preview' channel is invoked from AngularJs. The 'switch-channel' event  is sent from AngularJs with the '-preview' channel id, so the channel-store in ExtJs should be reloaded to update this change.