CHANNELMGR-584 cancel click simulation when cursor leaves component
Fixes the following edge case:

 1. User presses the mouse button on a component in a disabled
    container. The component won't move since its container is not
    registered with Dragula.
 2. User tries to move the component nonetheless, and moves the
    mouse cursor out of the component's element.
 3. Users gives up and release the mouse button.

In this case the drag-drop service still thinks the user is about
the click the component, since the mouseup handler on the
component's element is never called because the mouse button is
released outside the component's element. The only way then to 'fix'
the state is to click the same component again, which will trigger
a mouseup event on the component's element and show its component
properties. Meanwhile, other components cannot be dragged or
clicked anymore because the drag-drop service thinks we're already
dragging or clicking.

The fix is to register a mouseout handler on the component's
element when the drag-or-click begins. When the mouseout is
triggered and we are not dragging, we know the simulated click
should be canceled, and the internal state is reset.