CHANNELMGR-595 enabled 'view content in channel' again
Replaced the somewhat over-engineering redraw logic in
ChannelEditor.java with a more direct call to
ChannelEditor#loadChannel, which now gets the initial path as a
second parameter. The initial path is passed as a state parameter
to ChannelCtrl, which uses it to load the initial page. This
replaces the loading of the home page by the HippoIframeService.

The channelPrefix maintained by the ChannelService now includes
the mount path, so any initial path can be appended to it.