CHANNELMGR-596 show 'content' buttons (not functional yet)
The buttons are only visible in 'View' mode.

The PageStructureService now also parses comments of
HST-Type "CONTENT_LINK", and stores those in a contentLinks list.
After all content links have been parsed, the hippoIframe first
injects CSS that positions the content link's box elements before
generating the box elements themselves (the other way around
visually 'moves' the links, which is not very nice).

The markup of the content links is kept the same as before (an 'a'
tag with class 'hst-cmseditlink') so this API is kept backwards-
compatible (developers can use CSS to change the position of these
elements to position the buttons). The actual buttons are rendered
in the overlay, and their position is kept in sync the same way as
done with containers and components.