CHANNELMGR-563 don't prefix all render paths with the mount path
All render paths already include the mount path of the channel.
However, the 'channelPath' passed to the viewChannel method does
*not* include the mount path (i.e. it is a relative path within
the channel's URL space). So instead of prefixing all render paths
with the mount path, we now prefix the initial path with the mount
path. The 'initialPath' parameter of the 'channel' state has been
renamed to 'initialRenderPath' to reflect that it's a full render
path (including the mount path).