CHANNELMGR-597 reload page when channels perspective is activated
The current page is reloaded by sending a 'reload-page' event to
Angular, which gets picked up by the iframe service.

To prevent that 'view content in channel' (which also reloads the
page because the entire channel state is reloaded) also triggers a
second page reload because it also actives the channels perspective,
the iframe service now keeps track of whether a page has been
loaded. Until that has happened, reload-page events don't do
anything.

The 'show breadcrumb' code was called twice. To prevent that, the
RootPanel now sets a constructor property 'showBreadcrumbInitially'
when the channel manager perspective is active during the first page
load (e.g. a user pressed F5 when the channel manager was active).
Only in that case the RootPanel's constructor will show the
breadcrumb immediately.