CHANNELMGR-643 Use mouseleave to trigger component leave event
In Safari, the mouseout event is triggered for certain elements in
the site that handle mouse-pointer-events (like :hover or similar
events) when an overlay element is set to permeable. This means
that when a user starts dragging a component while hovering over
such an element, the mouse-out event is called which in turn
triggers our code to remove the 'permeable' class from the overlay
. Afterwards the onStartDrag event is called which adds the
'permeable' class again, but this is too late for Safari, i.e. it
will not pass pointer-events through the overlay element into the
site (where dragula takes over).

To fix this, I changed mouseout into mouseleave [1] which is only
triggered when the mouse pointer is moved off the element
(hovering nested elements, or in this case elements in an iframe
underneath the overlay element don't trigger this event).

There might be side-effects but AFAICS everything still works as
expected.

 [1] https://developer.mozilla.org/en-US/docs/Web/Events/mouseleave