CHANNELMGR-604 add required validation to channel settings
Added required validation and styling to string, picker (both path and image) and drop-down fields. The validation styling on drop-downs does not seem to be up to par with the styling for input fields: the dropdown needs to be "touched" or "dirty" before the feedback kicks in and the field turns red. I added no required support to boolean fields, because the UI widget can only differentiate between true and false, and not "absent".