CHANNELMGR-605 limit width of help tooltips and remove transparency
The width of the help icon tooltips is now limited to the sub-page
width (minus the width of the help icon itself, because the
tooltips are shown left of those help icons).

Improve readability of the tooltips by making the background color
black instead of the semi-transparent default color (which caused
the underlying widgets to shine through).