CHANNELMGR-701 Add logic to determine where to add a new item
When creating a new menu item, a marker item can be passed. If
this item has children, the new item will be created as the first
child node of the marker item. If the marker item has no children,
 the new item will be added as the next sibling of the marker item.

Adding a child menu item to an item without children is done by
using drag&drop.