CHANNELMGR-680 Fix creating preview config on first edit
The on-change method never called because it was a function reference
instead of a function call.

With that fixed, the toggle updates its model before calling the
on-change method, so the logic of setting isEditMode (possibly
asynchronously when preview config has to be created) no longer
worked.

Fixed by using a fake model for the toggle (because a model is
required by the ng-switch directive).