CHANNELMGR-715 Fixed test
On OSX (possibly on other platforms as well) it is possible to only show scrollbars when scrolling, in which case getScrollBarWidth return zero. And in this case zero is correct, as these scrollbars appear 'over' the site instead of next to it. So I changed the test to expect a number greater than -1.