CHANNELMGR-723 Add function isEmpty to check for child nodes
The builtin childNodes() method (from angular-ui-tree) is either buggy or not synced properly. A simple solution for this issue is to use a custom isEmpty function instead.

I also added a special purpose css class for nodes without children to add left-padding and allow them to be dragged by this white area (where previously a disabled, empty button would be rendered).