CHANNELMGR-817 Ensure paths always start with a slash
Usually the context path starts with a slash (e.g. '/site'), but
it can be empty when a channel's webapp is deployed as ROOT.war.
In that case the generated paths should still start with a slash,
otherwise the browser will resolve them relative to the Angular
app path instead of to the URL's hostname.

Since PathService.concatPaths already ignores the concatenation of
double slashes (e.g. '/' + '/site' = '/site') and it also ignores
falsy path elements (e.g. '/site' + undefined = '/site'), the
_makeContextPrefix logic could be simplified to a oneliner.