CHANNELMGR-813 Reload page when rendering new head contributions
The page structure service remembers all the head contributions
(processed and unprocessed) on the current page. When a container
or component is re-rendered, we check whether the new markup
also contains unprocessed head contributions that are not part
of the page yet. If so, we reload the page to make sure those
head contributions are also loaded.

When re-rendering components the page is only reloaded if no
custom property values are passed. Custom properties render
a 'preview' version of the component that is not saved in the
backend yet. Re-rendering the whole page would render the
saved component again instead of the preview version.