CHANNELMGR-896 Render 'multiple' fields
The values of each field are retrieved as an array to unify the code
for rendering single and multiple values. The label is only rendered
for the first value. Multiple fields with zero values are treated
separately because the regular loop over all values would not render
a label at all.

Animated labels would look strange for fields with multiple inputs,
hence none of the labels animates anymore by using md-no-float and
an empty placeholder value.

The height of the md-errors-spacer of subsequent multiple values
is set to zero to reduce the space between inputs of multiple fields.