CHANNELMGR-936 Don't reload page while loading an explicit one
A view-in-channel operation twiggers both a loadChannel call (to
load a specific page in a specific channel) and a reloadPage call
right after it (because the Channel Manager becomes active again).
Since the loadChannel call already reloads the page, the reloadPage
call is superfluous and logs a warning.

To prevent the warning, the selected channel is cleared until the
new channel has been set so the reloadPage call won't do anything.