CHANNELMGR-981 Only render .hst-cmseditlink when visible
An element is considered invisible if it or one of it's ancestors has the CSS style property visibility set to hidden, or it is not found with jQuery's :visible selector.

The #isVisible check should only be applied to hst-cmseditlink elements (edit document/menu buttons) and these should be actively hidden when the iframeElement is not considered visible.
Pass the jQuery element to avoid querying it again. First check
is(':visible') since that is the more common use case.