CHANNELMGR-966 Improve validation and error reporting on Update
Split up the writing logic into a writing part (throw exception if value doesn't match document type) and a validation part (limited to required validator for now). Upon Update, we push the values onto the draft variant, save to the repository, validate, and if all is well copy the changes to the preview variant.