CHANNELMGR-983 Determine iframe height via lowest element or marker
Using the document/body height requires a reset of the iframe's
height, which may trigger an endless loop if a channel uses a
resize listener that mutates the DOM.

Instead, we now use two different methods.

The default method is to calculate the bottom of the 'lowest' DOM
element in the page by traversing all elements in the body. The html
and body itself are skipped because they commonly get a height of
100%, which defeats the calculation. This method performs reasonably
well: for ~3500 nodes the worst case is around 120ms in IE11 on a VM
and ~20ms in Chrome.

The other method is to let projects mark an element as "the lowest
in the page" via the class hippo-channel-manager-page-height-marker.
If such a marker is found in the DOM we use its bottom as the page
height. This allows projects to circumvent the default method if it's
too slow or does not result in the correct height.

When html and body are set to 100%, a tall page could prevent a
not-so-tall next page from shrinking the height. To prevent this
the iframe height is always reset to 0 whenever a new page it loaded.

(cherry picked from commit 06a626a)