CHANNELMGR-996 Simplify height calculation
We don't need to take the existence of scrollbars into account for
determining the width and height of the channel. It just works.

The overlay width does not need to be reset since it'll be set at
the end of the function anyway (in setWidth). We do need the
iframe width reset, otherwise the getIframeWidth() calculation
won't work (the frame width will never shrink).