CHANNELMGR-923 Share pending changes logic between buttons
Both the 'publish' and 'open in full content' button need to save
pending changes (including a dialog) before opening the actual
content. The only difference is the mode in which the editor is
opened. Solution: reuse all logic and parameterize the mode.

Also combined the WicketExt events 'view-document' and
'edit-document' into a single 'open-document' event with the mode
as a parameter so the mode is consistently parameterized everywhere.