CHANNELMGR-898 Handle 'null' value as empty string
When the frontend uses an ngModel 'fieldValue.value', the posted
JSON is { value: 'somestring' } except when a required string
becomes empty. In that case the posted JSON becomes {} (an empty
object). This is default behavior of Angular. The value will then be
null in the backend. To avoid weird frontend constructs to work
around this behavior we treat a null value as a default value. For
strings the default is an empty string.