CHANNELMGR-898 Show one combined error message per multiple field
Multiple fields can have several inputs, each of which could be
invalid. As discussed with UX designer: instead of showing an error
message below each of the individual inputs (which would repeat the
error multiple times when multiple inputs are invalid) the errors
are combined into one error message that is shown below the multiple
field.