CHANNELMGR-1086 Extract ViewportService from OverlaySyncService
The logic to constrain the viewport will be independent of the
overlay. Once the iframe is resized by a constrained viewport, the
overlay should be redrawn automatically by a resize listener. The
debounce should not be necessary anymore, and has been removed.

The new viewport service only constrains the viewport. The scaling
service assumes the viewport is already set, and uses the current
width in the scaling calculation. The viewport toggle therefore
first updates the constrained width before syncing the scaling.

Also changed 'viewPort' to 'viewport' everywhere since it's one word.