CHANNELMGR-1086 Generate overlay in iframe
Renamed OverlaySyncService to OverlayService. It generates a div
with id 'hippo-overlay' in the iframe's body. The styling of the
overlay moved to hippo-iframe.css that was already injected into
the iframe. The buttons are not styled yet, they only show red boxes.

Generated box elements (for 'edit menu'/'edit content' links and for
empty containers) are now created by the page structure elements
instead of the PageStructureService (for the links) and the
OverlayElementController (for empty containers). This way the
structure element can keep track of whether the box element was
generated or not, which is needed later for the removal of the
generated box element.

Each box element (generated or not) gets an additional class
'hippo-overlay-box', which takes care of the min-height of 40px.

Combined 'edit menu' and 'edit content' links into 'embedded links'
array in the PageStructureService. Both links got their own type
so it's easy to switch in the page structure type in the
OverlaySyncService. The logic around each link type can move into
these classes later on.

The OverlayService registers an explicit changeListener with the
PageStructureService to get notified of changes. Alternatives seemed
suboptimal:
- letting the OverlayService use a watch on rootScope for changes
  in the page structure would not be efficient (called for every
  digest).
- letting the PageStructureService notify the OverlayService of
  changes directly would introduce a cyclic dependency.

TODO: remove overlay elements when removed from the
PageStructureService.