CHANNELMGR-1103 Fix cancellation of breadcrumb animation
Only the other delayed task was canceled, but the animation itself
could already be running. In that case the callback after the
old animation could be triggered after the callback of the new
animation. For example, a 'show' that canceled a previous 'hide'
could result in a hidden breadcrumb after all.

Fixed by animating via 'setWidth' with additional animation
options. ExtJs sets an 'anim' object in the options, and calling
anim.stop() cancels the animation. That way both the delayed task
and the animation itself are canceled before the next animation
starts.

Hiding/showing the toolbar element is not necessary since a
toolbar of 0px wide is also effectively hidden.

The animation time constant is not needed since 0.35 seconds is
the default animation time of ExtJs.