CHANNELMGR-1101 Prevent double transitionend callback in Firefox
Firefox calls the transitionend callback directly again when
transform is modified again inside the transitionend callback, which
results in calling scrollBy twice. To avoid this, we remove the
transitionend callback directly after it has been invoked for the
first time.

Also fixed the reset of 'this.animating' in the callback. Since
'this' in a jQuery callback always refers to an element, we have
to use self = this to be able to modify the service's field.