CHANNELMGR-1093 Reorder components via drag-drop
Register a mousedown click handler on the overlay root element. On mouse down, check if a component element is clicked and check if a component-mouse-down handler is registered (which is done when edit-mode is enabled). Delegate click to the dragDropService which uses the existing behavior to decide if it is a click (open properties window) or a drag-drop action.

Visibility of overlay-elements is controlled by CSS classes on the root (HTML) element, all overlay-elements are hidden by default.