CHANNELMGR-1130 Set transform-origin of html via .hippo-scale class
Using a separate CSS class for the transform-origin is cleaner
than setting it directly on the HTML element. It also allows
custom styling of the page when it is scaled down (may that need
ever arise in projects).

Also remove the 'transform: scale' style when the page is not
scaled anymore to lower the possible impact of the channel manager
styling (e.g. overriding custom transform's on the HTMl element).

Simplified the 'expect' function in the test a little since they
only get one element.