CHANNELMGR-1130 Glue mirror container on top of the iframe
The container for the Dragula mirror should have the same size and
position as the iframe because the coordinates of the mirror are
relative to the iframe's viewport. When the mirror container in
the app has a different size or position the mirror will be
positioned in the wrong place. That which was the case in, for
example, 'Tablet' and 'Mobile' view, because the iframe was
shifted to the right but the mirror container was not.

The viewport service now restricts the width of the sheet. The
scaling service shifts the sheet but still scales the iframe.

The iframe and mirror container both take 100% width and height of
the sheet so they are positioned on top of each other. The
'position: absolute' of the mirror container ensures that the
left and top of the mirror are relative to the mirror container.